Refactor async server handlers into classes
