+++ tests/Ajax/TimeoutStatusTest.php
@@ -10,38 +10,11 @@ namespace {
         }
     }
 
-    if (!function_exists('timeout_status')) {
-        function timeout_status($args = false): \Jaxon\Response\Response
-        {
-            global $session, $start_timeout_show_seconds, $never_timeout_if_browser_open, $settings;
-
-            $response = \Jaxon\jaxon()->newResponse();
-            if ($args === false || !isset($session['user'])) {
-                return $response;
-            }
-
-            $lastOn = strtotime($session['user']['laston']);
-            $elapsed = time() - $lastOn;
-            $remaining = $settings->getSetting('LOGINTIMEOUT', 900) - $elapsed;
-
-            $warning = '';
-            if ($remaining <= 0) {
-                $warning = 'Your session has timed out!';
-            } elseif ($remaining < $start_timeout_show_seconds) {
-                $warning = 'TIMEOUT';
-            }
-
-            if ($warning !== '') {
-                $response->assign('notify', 'innerHTML', $warning);
-            }
-
-            return $response;
-        }
-    }
 }
 
 namespace Lotgd\Tests\Ajax {
 
+    use Lotgd\Async\Handler\Timeout;
     use PHPUnit\Framework\TestCase;
 
     final class TimeoutStatusTest extends TestCase
@@ -67,13 +40,13 @@ namespace Lotgd\Tests\Ajax {
                 }
             };
           
-            require_once __DIR__ . '/../../async/server.php';
+            require_once __DIR__ . '/../bootstrap.php';
 
         }
 
         public function testTimeoutWarningIsReturned(): void
         {
-            $response = \timeout_status(true);
+            $response = (new Timeout())->timeoutStatus(true);
             $commands = $response->getCommands();
 
             $this->assertNotEmpty($commands);
@@ -86,7 +59,7 @@ namespace Lotgd\Tests\Ajax {
             global $session;
             $session = [];
 
-            $response = \timeout_status(true);
+            $response = (new Timeout())->timeoutStatus(true);
             $this->assertEmpty($response->getCommands());
         }
     }
